LayoutTests/fast/viewport/viewport-39-expected.txt

ALERT: viewport size 200x700 scale 1.6 with limits [1.6, 5] and userScalable true
